[E2E] Revert running all `question` repros together (#22107)
