--- frontend/test/metabase/scenarios/question/question-reproductions.cy.spec.js
@@ -1,43 +1,0 @@
-import { issue4482 } from "./reproductions/4482-temporal-min-max";
-import { issue6239 } from "./reproductions/6239-sort-using-cust-exp";
-import { issue9027 } from "./reproductions/9027-new-questions-not-in-saved-questions-immediately";
-import { issue13097 } from "./reproductions/13097-mongo-apply-distinct-count-multiple-columns";
-import { issue13263 } from "./reproductions/13263-postgres-show-row-details-on-pk-click";
-import { issue14957 } from "./reproductions/14957-unable-to-save-question-before-query-executed";
-import { issue15714 } from "./reproductions/15714-cc-postgres-percentile-accepts-two-params";
-import { issue15876 } from "./reproductions/15876-postgres-cast-time";
-import { issue16621 } from "./reproductions/16621-create-multiple-filters-with-same-value";
-import { issue16938 } from "./reproductions/16938-nested-native-query-pk-drill";
-import { issue17512 } from "./reproductions/17512";
-import { issue17514 } from "./reproductions/17514-ui-overlay";
-import { issue17963 } from "./reproductions/17963-mongo-filter-expression-compare-two-fields";
-import { issue18207 } from "./reproductions/18207-string-min-max";
-import { issue18978 } from "./reproductions/18978-18977-nested-question-nodata-user";
-import { issue19341 } from "./reproductions/19341-disabled-nested-queries";
-import { issue19742 } from "./reproductions/19742-data-picker-closes-after-hiding-table";
-import { issue20551 } from "./reproductions/20551-filter-starts-with";
-import { issue20627 } from "./reproductions/20627-nested-long-names-wrong-aliases";
-import { issue20683 } from "./reproductions/20683-postgres-current-quarter";
-import { issue17910 } from "./reproductions/17910-revision-history-update";
-
-issue4482();
-issue6239();
-issue9027();
-issue13097();
-issue13263();
-issue14957();
-issue15714();
-issue15876();
-issue16621();
-issue16938();
-issue17512();
-issue17514();
-issue17963();
-issue18207();
-issue18978();
-issue19341();
-issue19742();
-issue20551();
-issue20627();
-issue20683();
-issue17910();
