+++ frontend/test/metabase/scenarios/question/reproductions/16621-create-multiple-filters-with-same-value.cy.spec.js
@@ -1,0 +1,21 @@
+import { restore, openProductsTable } from "__support__/e2e/cypress";
+
+describe("issue 16661", () => {
+  beforeEach(() => {
+    restore();
+    cy.signInAsAdmin();
+    openProductsTable({ limit: 3 });
+  });
+
+  it("should be possible to create multiple filter that start with the same value (metabase#16621)", () => {
+    cy.findByText("Category").click();
+    cy.findByText("Filter by this column").click();
+    cy.findByPlaceholderText("Search the list").type("Doo{enter}");
+    cy.findByTestId("Doo-filter-value").click();
+    cy.button("Add filter").click();
+    cy.findByText("Category is Doo").click();
+    cy.findByTestId("Doohickey-filter-value").click();
+    cy.button("Update filter").click();
+    cy.findByText("Category is 2 selections");
+  });
+});
